+++ scripts/benchmarks.sh
@@ -1,6 +1,6 @@
 num_parallel=24
-num_seeds=24
 num_gpus=0
+num_seeds=24
 seed_offset=0 # $num_seeds
 
 datasets=("ihdp-N" "news-N" "sim-N")
@@ -25,18 +25,18 @@ do
             fi
             flags="${extra_flags}"
             (
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir aipw";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "aipw" ${flags};
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir tresnet --tr_reg";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "tresnet" --tr_reg ${flags};
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir plugin";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "plugin" --outcome_only ${flags};
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir IPW";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "ipw" --ratio_only ${flags};
-               echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir drnet";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "drnet" --outcome_only --drnet ${flags};
-               echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir drnet --tr_reg";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "drnet_tr" --tr_reg --drnet ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir aipw";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "aipw" ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir tresnet --tr_reg";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "tresnet" --tr_reg ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir plugin";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "plugin" --outcome_only ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir IPW";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "ipw" --ratio_only ${flags};
+               echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir drnet";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "drnet" --outcome_only --drnet ${flags};
+               echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir drnet --tr_reg";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "drnet_tr" --tr_reg --drnet ${flags};
             ) &
         done
         wait
@@ -62,10 +62,10 @@ do
             fi
             flags="${extra_flags}"
             (
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir c_ratio --tr_reg --ratio c_ratio";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "c_ratio" --ratio c_ratio --tr_reg ${flags};
-                echo "Running --dataset ${dset} main_tres.py --seed=$s --rdir results/benchmarks --edir erm --tr_reg --ratio erm";
-                python main_tres.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "erm" --ratio erm --tr_reg ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir c_ratio --tr_reg --ratio c_ratio";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "c_ratio" --ratio c_ratio --tr_reg ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$s --rdir results/benchmarks --edir erm --tr_reg --ratio erm";
+                python main.py --dataset ${dset} --seed=$s --rdir "results/benchmarks" --edir "erm" --ratio erm --tr_reg ${flags};
             ) &
         done
         wait
@@ -91,10 +91,10 @@ do
             fi
             flags="${extra_flags}"
             (
-                echo "Running --dataset ${dset} main_tres.py --seed=$((s + seed_offset)) --rdir results/benchmarks --edir non_poisson";
-                python main_tres.py --dataset ${dset} --seed=$((s + seed_offset)) --rdir "results/benchmarks" --edir "non_poisson" --count --tr_reg ${flags};
-                echo "Running --dataset ${dset} main_tres.py --seed=$((s + seed_offset)) --rdir results/benchmarks --edir poisson";
-                python main_tres.py --dataset ${dset} --seed=$((s + seed_offset)) --rdir "results/benchmarks" --edir "poisson" --poisson --count --tr_reg ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$((s + seed_offset)) --rdir results/benchmarks --edir non_poisson";
+                python main.py --dataset ${dset} --seed=$((s + seed_offset)) --rdir "results/benchmarks" --edir "non_poisson" --count --tr_reg ${flags};
+                echo "Running --dataset ${dset} main.py --seed=$((s + seed_offset)) --rdir results/benchmarks --edir poisson";
+                python main.py --dataset ${dset} --seed=$((s + seed_offset)) --rdir "results/benchmarks" --edir "poisson" --poisson --count --tr_reg ${flags};
             ) &
         done
         wait
